requirements.txt
asgiref==3.8.1
asttokens==3.0.0
attrs==23.2.0
beautifulsoup4==4.13.3
branca==0.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.9
decorator==5.1.1
Django==4.2
django-admin-interface==0.29.4
django-colorfield==0.12.0
django-easy-select2==1.5.8
django-flat-responsive==2.0
dnspython==2.6.1
executing==2.1.0
fiona==1.9.6
flake8==7.0.0
folium==0.16.0
fonttools==4.55.0
geopandas==0.14.4
gunicorn==23.0.0
idna==3.7
ipykernel==6.29.5
ipython==8.30.0
ipywidgets==8.1.5
jedi==0.19.2
Jinja2==3.1.4
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.7.1
matplotlib-inline==0.1.7
mccabe==0.7.0
nest-asyncio==1.6.0
numpy==1.26.4
packaging==24.0
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.1.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pycodestyle==2.11.1
pydeck==0.7.1
pyflakes==3.2.0
Pygments==2.18.0
pymongo==4.3.3
pyparsing==3.2.0
pyproj==3.6.1
pyshp==2.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.1
pyzmq==26.2.0
requests==2.32.3
shapely==2.0.1
six==1.16.0
soupsieve==2.6
sqlparse==0.5.0
stack-data==0.6.3
text-unidecode==1.3
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
widgetsnbextension==4.0.13
xyzservices==2024.4.0
